+++ server/TWS Foundation/TWS Foundation/Controllers/Business/TrailersController.cs
@@ -21,4 +21,13 @@ public class TrailersController : ControllerBase {
     public async Task<IActionResult> View(SetViewOptions<Trailer> Options) {
         return Ok(await Service.View(Options));
     }
+    [HttpPost(), Auth([])]
+    public async Task<IActionResult> Create(Trailer[] Trailers) {
+        return Ok(await Service.Create(Trailers));
+    }
+
+    [HttpPost(), Auth([])]
+    public async Task<IActionResult> Update(Trailer Trailer) {
+        return Ok(await Service.Update(Trailer));
+    }
 }