+++ server/TWS Foundation/TWS Foundation/Controllers/Business/TrailersExternalsController.cs
@@ -21,4 +21,14 @@ public class TrailersExternalsController : ControllerBase {
     public async Task<IActionResult> View(SetViewOptions<TrailerExternal> Options) {
         return Ok(await Service.View(Options));
     }
+
+    [HttpPost(), Auth([])]
+    public async Task<IActionResult> Create(TrailerExternal[] Trailers) {
+        return Ok(await Service.Create(Trailers));
+    }
+
+    [HttpPost(), Auth([])]
+    public async Task<IActionResult> Update(TrailerExternal Trailer) {
+        return Ok(await Service.Update(Trailer));
+    }
 }