+++ server/TWS Foundation/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## CURRENT - [xx.xx-xxxx]
 
 - Notes:
-
+    1. [Trailers] and [TrailersExternals] added the Create and Update depot services.  
+    1. 
 - Dependencies:
 
 ## 6.0.0 - [03.11-2024]